#139 - Refactoring tests
